pyproject.toml
[tool.poetry]
name = "gcloud-pubsub-emulador"
version = "0.1.0"
description = ""
authors = ["Wanderson Silva <bsilva.wanderson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-pubsub = "^2.13.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
